The machine language(×64
intel) byte length of the VM loop method was reduced.
It may slow down in some cases, but overall this speeded up the process.
| MethodSize (byte)| 
|------------ |
| 28940  | 
| 20171| 

## Loop Bench
```lua
a = 0.1
while a < 100000
  a += 1
  a -= 0.1
  a *= 1.001
  a /= 1.001
end
```
| Method      | Mean     | Error     | StdDev    | Allocated |
|------------ |---------:|----------:|----------:|----------:|
| MRubyCS     | 3.084 ms | 0.0158 ms | 0.0105 ms |     176 B |
| MRubyCS(PR)  | 2.661 ms | 0.0035 ms | 0.0018 ms |     176 B |
 MRubyNative | 4.393 ms | 0.0186 ms | 0.0123 ms |       1 B |

## Fib Bench
```lua
def fib n
  return n if n < 2
  fib(n-2) + fib(n-1)
end
fib(20)
```
| Method      | Mean       | Error    | StdDev  | Allocated |
|------------ |-----------:|---------:|--------:|----------:|
| MRubyCS     | 2,051.7 us | 9.65 us | 6.38 us |     354 B |
| MRubyCS(PR) | 1,897.8 us | 10.16 us | 6.72 us |     353 B |
| MRubyNative |   576.6 us |  1.49 us | 0.99 us |         - |